+++ openeo/rest/connection.py
@@ -66,6 +66,11 @@ from openeo.rest.datacube import DataCube, InputDate
 from openeo.rest.graph_building import CollectionProperty
 from openeo.rest.job import BatchJob, RESTJob
 from openeo.rest.mlmodel import MlModel
+from openeo.rest.models.general import (
+    CollectionListingResponse,
+    JobListingResponse,
+    ProcessListingResponse,
+)
 from openeo.rest.service import Service
 from openeo.rest.udp import Parameter, RESTUserDefinedProcess
 from openeo.rest.userfile import UserFile
@@ -671,7 +676,7 @@ class Connection(RestApiConnection):
     def user_jobs(self) -> List[dict]:
         return self.list_jobs()
 
-    def list_collections(self) -> List[dict]:
+    def list_collections(self) -> CollectionListingResponse:
         """
         List basic metadata of all collections provided by the back-end.
 
@@ -682,10 +687,15 @@ class Connection(RestApiConnection):
             it is recommended to use :py:meth:`~openeo.rest.connection.Connection.describe_collection` instead.
 
         :return: list of dictionaries with basic collection metadata.
+
+        .. versionchanged:: 0.38.0
+            Returns a :py:class:`~openeo.rest.models.general.CollectionListingResponse` object
+            instead of a simple ``List[dict]``.
         """
         # TODO: add caching #383, but reset cache on auth change #254
-        data = self.get('/collections', expected_status=200).json()["collections"]
-        return VisualList("collections", data=data)
+        # TODO #677 add pagination support?
+        data = self.get("/collections", expected_status=200).json()
+        return CollectionListingResponse(response_data=data)
 
     def list_collection_ids(self) -> List[str]:
         """
@@ -822,23 +832,27 @@ class Connection(RestApiConnection):
         # TODO: duplication with `Connection.describe_collection`: deprecate one or the other?
         return CollectionMetadata(metadata=self.describe_collection(name))
 
-    def list_processes(self, namespace: Optional[str] = None) -> List[dict]:
-        # TODO: Maybe format the result dictionary so that the process_id is the key of the dictionary.
+    def list_processes(self, namespace: Optional[str] = None) -> ProcessListingResponse:
         """
         Loads all available processes of the back end.
 
         :param namespace: The namespace for which to list processes.
 
-        :return: processes_dict: Dict All available processes of the back end.
+        :return: listing of available processes
+
+        .. versionchanged:: 0.38.0
+            Returns a :py:class:`~openeo.rest.models.general.ProcessListingResponse` object
+            instead of a simple ``List[dict]``.
         """
+        # TODO: Maybe format the result dictionary so that the process_id is the key of the dictionary.
+        # TODO #677 add pagination support?
         if namespace is None:
-            processes = self._capabilities_cache.get(
-                key=("processes", "backend"),
-                load=lambda: self.get('/processes', expected_status=200).json()["processes"]
+            response = self._capabilities_cache.get(
+                key=("processes", "backend"), load=lambda: self.get("/processes", expected_status=200).json()
             )
         else:
-            processes = self.get('/processes/' + namespace, expected_status=200).json()["processes"]
-        return VisualList("processes", data=processes, parameters={'show-graph': True, 'provide-download': False})
+            response = self.get("/processes/" + namespace, expected_status=200).json()
+        return ProcessListingResponse(response_data=response)
 
     def describe_process(self, id: str, namespace: Optional[str] = None) -> dict:
         """
@@ -857,26 +871,25 @@ class Connection(RestApiConnection):
 
         raise OpenEoClientException("Process does not exist.")
 
-    def list_jobs(self, limit: Union[int, None] = None) -> List[dict]:
+    def list_jobs(self, limit: Union[int, None] = None) -> JobListingResponse:
         """
-        Lists all jobs of the authenticated user.
+        Lists (batch) jobs metadata of the authenticated user.
 
         :param limit: maximum number of jobs to return. Setting this limit enables pagination.
 
         :return: job_list: Dict of all jobs of the user.
 
-        .. versionadded:: 0.36.0
+        .. versionchanged:: 0.36.0
             Added ``limit`` argument
+
+        .. versionchanged:: 0.38.0
+            Returns a :py:class:`~openeo.rest.models.general.JobListingResponse` object
+            instead of simple ``List[dict]``.
         """
         # TODO: Parse the result so that Job classes returned?
-        resp = self.get("/jobs", params={"limit": limit}, expected_status=200).json()
-        if resp.get("federation:missing"):
-            _log.warning("Partial user job listing due to missing federation components: {c}".format(
-                c=",".join(resp["federation:missing"])
-            ))
         # TODO: when pagination is enabled: how to expose link to next page?
-        jobs = resp["jobs"]
-        return VisualList("data-table", data=jobs, parameters={'columns': 'jobs'})
+        resp = self.get("/jobs", params={"limit": limit}, expected_status=200).json()
+        return JobListingResponse(response_data=resp)
 
     def assert_user_defined_process_support(self):
         """
@@ -928,13 +941,18 @@ class Connection(RestApiConnection):
         )
         return udp
 
-    def list_user_defined_processes(self) -> List[dict]:
+    def list_user_defined_processes(self) -> ProcessListingResponse:
         """
         Lists all user-defined processes of the authenticated user.
+
+        .. versionchanged:: 0.38.0
+            Returns a :py:class:`~openeo.rest.models.general.ProcessListingResponse` object
+            instead of a simple ``List[dict]``.
         """
+        # TODO #677 add pagination support?
         self.assert_user_defined_process_support()
-        data = self.get("/process_graphs", expected_status=200).json()["processes"]
-        return VisualList("processes", data=data, parameters={'show-graph': True, 'provide-download': False})
+        data = self.get("/process_graphs", expected_status=200).json()
+        return ProcessListingResponse(response_data=data)
 
     def user_defined_process(self, user_defined_process_id: str) -> RESTUserDefinedProcess:
         """